feat(configureView(new), status, xmake): Add new view (#304)
* ✨ feat(configureView(new), status, xmake): Add new view

Based on the current status management, a tree display page of project configuration has been added to the Activity Bar section. Since there is no decoupling of Status and StatusBarItem, try to read the information by adding a getter while reflesh when calling the setter to keep the data in sync.

https://github.com/xmake-io/xmake-vscode/issues/303

* Update xmake.ts

---------

Co-authored-by: ruki <waruqi@gmail.com>